refactor(server): person dto
