--- server/src/domain/person/dto/index.ts
@@ -1,1 +1,0 @@
-export * from './person-update.dto';
