--- server/src/domain/person/dto/person-update.dto.ts
@@ -1,7 +1,0 @@
-import { IsNotEmpty, IsString } from 'class-validator';
-
-export class PersonUpdateDto {
-  @IsNotEmpty()
-  @IsString()
-  name!: string;
-}
