+++ server/src/domain/person/index.ts
@@ -1,4 +1,3 @@
-export * from './dto';
+export * from './person.dto';
 export * from './person.repository';
 export * from './person.service';
-export * from './response-dto';
