+++ server/src/domain/person/person.dto.ts
@@ -1,4 +1,11 @@
 import { AssetFaceEntity, PersonEntity } from '@app/infra/entities';
+import { IsNotEmpty, IsString } from 'class-validator';
+
+export class PersonUpdateDto {
+  @IsNotEmpty()
+  @IsString()
+  name!: string;
+}
 
 export class PersonResponseDto {
   id!: string;
