--- server/src/domain/person/response-dto/index.ts
@@ -1,1 +1,0 @@
-export * from './person-response.dto';
